+++ src/VivViewer.js
@@ -81,9 +81,7 @@ export default class VivViewer extends PureComponent {
           ...overview,
           viewWidth,
           viewHeight,
-          loader,
-          margin: overview.margin,
-          overviewScale: overview.scale
+          loader
         });
         // eslint-disable-next-line react/no-did-update-set-state
         this.setState(prevState => {
@@ -139,16 +137,15 @@ export default class VivViewer extends PureComponent {
   }
 
   render() {
-    /* eslint-disable react/destructuring-assignment */
-    const { loader } = this.props;
-    const { overview } = this.state;
+    const { loader, viewHeight, viewWidth } = this.props;
+    const { overview, viewState } = this.state;
 
     const views = [
       new OrthographicView({
         id: 'detail',
         controller: true,
-        height: this.props.viewHeight,
-        width: this.props.viewWidth
+        height: viewHeight,
+        width: viewWidth
       })
     ];
 
@@ -162,7 +159,7 @@ export default class VivViewer extends PureComponent {
         layers={this._renderLayers()}
         onViewStateChange={this._onViewStateChange}
         views={views}
-        viewState={this.state.viewState}
+        viewState={viewState}
       />
     ) : (
       <div className="viv-error">
@@ -172,6 +169,5 @@ export default class VivViewer extends PureComponent {
         </p>
       </div>
     );
-    /* eslint-disable react/destructuring-assignment */
   }
 }