+++ src/index.js
@@ -1,6 +1,6 @@
 import { VivViewerLayer, StaticImageLayer } from './layers';
 import VivViewer from './VivViewer';
-import { createTiffPyramid, createZarrPyramid, ZarrLoader } from './loaders';
+import { createTiffPyramid, createZarrLoader, ZarrLoader } from './loaders';
 
 export {
   VivViewerLayer,
@@ -8,5 +8,5 @@ export {
   StaticImageLayer,
   ZarrLoader,
   createTiffPyramid,
-  createZarrPyramid
+  createZarrLoader
 };